#  Add UTf8Json library to JsonWriter performance benchmarks